% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mapped_function.R
\name{get_mapped_function}
\alias{get_mapped_function}
\alias{get_mapped_function_name}
\title{Retrieve the function applied to each column pair}
\usage{
get_mapped_function(data)

get_mapped_function_name(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Function or string (length 1).
}
\description{
If you pass a function \code{f()} to \code{as_colpair_mapper()}, and then
call the resulting function, you will get a data frame back. Here are two
functions to retrieve \code{f()} from that data frame:
\itemize{
\item \code{get_mapped_function()} returns the function itself.
\item \code{get_mapped_function_name()} returns the function's name as a string.
}
}
\examples{
# `covary()` was made by `as_colpair_mapper()`:
df <- covary(mtcars)

# Since `covary()` applies `stats::cov()`,
# this is the function returned here:
cov_from_df <- get_mapped_function(data = df)
identical(cov_from_df, stats::cov)

# Just get the name instead:
get_mapped_function_name(data = df)
}
