% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pairmaps_df.R
\name{is_pairmaps_df}
\alias{is_pairmaps_df}
\title{Is an object a pairmaps data frame?}
\usage{
is_pairmaps_df(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A single Boolean value.
}
\description{
\code{is_pairmaps_df()} tests if an object is a data frame returned
by a function that was made by \code{as_colpair_mapper()}.
}
\details{
The function checks whether \code{x} is a data frame that inherits
exactly one S3 class that starts on \code{"pairmaps_df_"}. This is a given if
\code{x} was returned by a function created by \code{as_colpair_mapper()} and its
classes have not been manipulated.
}
\examples{
# Basic function, from `?corrr::colpair_map()`:
calc_p_value <- function(vec_a, vec_b) {
  stats::t.test(vec_a, vec_b)$p.value
}

# Produce a new mapper function:
p_value_map <- as_colpair_mapper(calc_p_value)

is_pairmaps_df(p_value_map(mtcars))

is_pairmaps_df(mtcars)

is_pairmaps_df(4)
}
