% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vary-covary.R
\name{vary-covary}
\alias{vary-covary}
\alias{vary}
\alias{covary}
\title{Variance and covariance data frames}
\usage{
vary(.data, ..., .diagonal = NA, .quiet = FALSE)

covary(.data, ..., .diagonal = NA, .quiet = FALSE)
}
\arguments{
\item{.data}{Numeric vector, matrix or data frame.}

\item{...}{Further arguments passed on to \code{var()} or \code{cov()}, respectively.}

\item{.diagonal}{Value to which the diagonal will be set. Default is \code{NA}.}

\item{.quiet}{Boolean. Set it to \code{TRUE} to suppress the message about
function mapping.}
}
\value{
Both functions return \code{cor_df} tibbles, i.e., "correlation data
frames". However, they don't compute correlations: This name simply
reflects the origins of the \code{cor_df} class in the corrr package.
}
\description{
\code{vary()} and \code{covary()} return variance and covariance data
frames instead of matrices, making them easier to explore with the \href{https://corrr.tidymodels.org/articles/using-corrr.html}{corrr package}.
}
\examples{
vary(mtcars)

covary(mtcars)
}
\seealso{
\code{as_colpair_mapper()}, which made these functions via
\code{corrr::colpair_map()}.

\code{vary()} wraps \code{stats::var()}, \code{covary()} wraps \code{stats::cov()}. For
the analogous \code{stats::cor()} wrapper, see \code{corrr::correlate()}.
}
